succeed to work properly in webgpu mode

+++ src/webgpu/shaderity_shaders/nodes/WorldMatrix.vert.wgsl
@@ -1,4 +1,4 @@
 
 fn worldMatrix(outValue: ptr<function, mat4x4<f32>>) {
-  *outValue = get_worldMatrix(a_instanceIds.x);
+  *outValue = get_worldMatrix(u32(a_instanceIds.x));
 }
